requirements.txt
streamlit
python-dotenv
pypdf
faiss-cpu
openai
langchain-nvidia-ai-endpoints
langchain-community
